Add a link to the product page from the catalogue
The first part of the test was easy to implement, we just need to add a link to the product page.

Note the use of {{ route('product.show', ..) }}

This is in line with wishful programming, I don't want to have to hard code urls in my code, and couple the endpoints with my templates.

This would be a nice way to write what we want to achieve.