Add User integration test
No we're demonstrating the outside-in approach. We begun with a failing acceptance test, which is still red.

To move further along in the test, we need to start testing at a lower level, as our Controller is now defining logic on the User class, which doesn't exist.

To build that feature, we need a test.